requirements.txt

# packages
conda
pytorch[>=1.6.0]
ase
dscribe
h5py

# data
skf data for training
hdf dataset for training
